requirements.txt
torch
pandas
Pillow
sentence-transformers
git+https://github.com/openai/CLIP.git
